[ENC] PR comments - Set up RustWarden struct
