+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "RustWarden"
+name = "rust_warden"
 version = "0.1.0"
 edition = "2024"
