+++ src/commands/list_interfaces.rs
@@ -1,7 +1,8 @@
 use pnet::datalink::interfaces;
 
-pub fn list_interfaces() {
+pub fn list_interfaces() -> Result<(), String> {
     for interface in interfaces() {
         println!("{}", interface);
     }
+    Ok(())
 }
